report

+++ simplenns/cifar10_fc.py
@@ -116,7 +116,7 @@ def inference(images):
 
 def correct_ones(logits, labels):
     correct_prediction = tf.nn.in_top_k(logits, labels, k=1)
-    correct_prediction = tf.cast(correct_prediction, tf.float32)
+    correct_prediction = tf.cast(correct_prediction, tf.int32)
     # pred = tf.cast(tf.nn.in_top_k(logits, labels, k=1), tf.float32)
     n_corrects = tf.reduce_sum(correct_prediction,name='correct_n')
     tf.summary.scalar('correct_ones', n_corrects)
